backend/requirements.txt
fastapi
pymongo
uvicorn
python-multipart
passlib
PyJWT
langchain
groq
python-dotenv
plotly
matplotlib
pandas
sqlalchemy
pymysql
psycopg2-binary
openpyxl
cryptographyhttpx
bcrypt